requirements.txt
azure-identity==1.14.0
Flask==2.3.2
openai~=1.3.2
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0